+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Verb Changelog
+## master
+- Bind `verb-export-request-on-point-verb` to `C-b` in the Verb command map.
+- Bind `verb-export-request-on-point-human` to `C-n` in the Verb command map.
+
 ## **2.6.0** - 2020-02-03
 - Removed `verb-max-redirections` variable, as it wasn't working as intended. Use `url-max-redirections` instead.
 - Add `default` argument to `verb-var`.